+++ e2e_test/iceberg/main.py
@@ -1,252 +1,28 @@
 import argparse
 import configparser
-import decimal
-import glob
 import os
-import subprocess
 import sys
-import time
-import unittest
-from datetime import date, datetime, timedelta, timezone
-from enum import Enum, auto
-from typing import Any, Dict, List, NamedTuple, Optional, Tuple
+from datetime import datetime
+from typing import Optional
 
 import tomli as toml
-from pyspark.sql import SparkSession
+
+from test_utils import (
+    log,
+    LogLevel,
+    TestResult,
+    print_test_summary,
+    init_iceberg_table,
+    compare_sql,
+    drop_table,
+    discover_test_cases,
+    execute_slt,
+    prepare_test_env,
+    verify_result,
+)
 
 
-class LogLevel(Enum):
-    DEBUG = auto()
-    INFO = auto()
-    WARNING = auto()
-    ERROR = auto()
-    HEADER = auto()
-    SEPARATOR = auto()
-    RESULT = auto()
-
-
-class Color:
-    HEADER = "\033[95m"
-    BLUE = "\033[94m"
-    GREEN = "\033[92m"
-    YELLOW = "\033[93m"
-    RED = "\033[91m"
-    ENDC = "\033[0m"
-    BOLD = "\033[1m"
-    UNDERLINE = "\033[4m"
-
-
-def log(msg, level: LogLevel = LogLevel.INFO, indent: int = 0):
-    """Enhanced logging function with color support and structure.
-
-    Args:
-        msg: The message to log
-        level: LogLevel enum indicating the type of message
-        indent: Number of spaces to indent the message
-    """
-    prefix = " " * indent
-    timestamp = datetime.now().strftime("%H:%M:%S")
-
-    if level == LogLevel.DEBUG:
-        color = Color.BLUE
-        prefix = f"{prefix}[DEBUG] "
-    elif level == LogLevel.INFO:
-        color = Color.GREEN
-        prefix = f"{prefix}[INFO] "
-    elif level == LogLevel.WARNING:
-        color = Color.YELLOW
-        prefix = f"{prefix}[WARN] "
-    elif level == LogLevel.ERROR:
-        color = Color.RED
-        prefix = f"{prefix}[ERROR] "
-    elif level == LogLevel.HEADER:
-        color = Color.HEADER + Color.BOLD
-        prefix = f"{prefix}=== "
-        msg = f"{msg} ==="
-    elif level == LogLevel.SEPARATOR:
-        color = Color.BLUE
-        msg = "=" * 40
-        prefix = prefix
-    elif level == LogLevel.RESULT:
-        color = Color.GREEN + Color.BOLD
-        prefix = f"{prefix}>>> "
-    else:
-        color = ""
-        prefix = prefix
-
-    # Only use colors if we're writing to a terminal
-    if not sys.stdout.isatty():
-        color = ""
-        Color.ENDC = ""
-
-    print(f"{color}{prefix}{msg}{Color.ENDC}", flush=True)
-
-
-def strtobool(v):
-    return v.lower() == "true"
-
-
-def strtodate(v):
-    return date.fromisoformat(v)
-
-
-def strtots(v):
-    return datetime.fromisoformat(v).astimezone(timezone.utc).replace(tzinfo=None)
-
-
-g_spark: Optional[SparkSession] = None
-
-
-def get_spark(args) -> SparkSession:
-    spark_config = args["spark"]
-    global g_spark
-    if g_spark is None:
-        g_spark = SparkSession.builder.remote(spark_config["url"]).getOrCreate()
-
-    return g_spark
-
-
-def init_iceberg_table(args, init_sqls):
-    log(f"init_sqls:", level=LogLevel.INFO)
-    spark = get_spark(args)
-    for sql in init_sqls:
-        log(sql, level=LogLevel.INFO, indent=2)
-        spark.sql(sql)
-
-
-def execute_slt(args, slt):
-    if slt is None or slt == "":
-        return
-    rw_config = args["risingwave"]
-    cmd = f"sqllogictest -p {rw_config['port']} -d {rw_config['db']} {slt}"
-    log(f"Executing slt: {slt}", level=LogLevel.INFO)
-    subprocess.run(cmd, shell=True, check=True)
-    time.sleep(15)
-
-
-def verify_result(args, verify_sql, verify_schema, verify_data):
-    tc = unittest.TestCase()
-
-    time.sleep(3)
-    log(f"verify_result:", level=LogLevel.HEADER)
-    log(f"Executing sql: {verify_sql}", level=LogLevel.INFO, indent=2)
-    spark = get_spark(args)
-    df = spark.sql(verify_sql).collect()
-    log(f"Result:", level=LogLevel.HEADER, indent=2)
-    log("", level=LogLevel.SEPARATOR, indent=2)
-    for row in df:
-        log(row, level=LogLevel.RESULT, indent=4)
-    log("", level=LogLevel.SEPARATOR, indent=2)
-    rows = verify_data.splitlines()
-    tc.assertEqual(len(df), len(rows), "row length mismatch")
-    tc.assertEqual(len(verify_schema), len(df[0]), "column length mismatch")
-    for row1, row2 in zip(df, rows):
-        log(f"Row1: {row1}, Row 2: {row2}", level=LogLevel.DEBUG, indent=4)
-        # New parsing logic for row2
-        row2 = parse_row(row2)
-        for idx, ty in enumerate(verify_schema):
-            if ty == "int" or ty == "long":
-                tc.assertEqual(row1[idx], int(row2[idx]))
-            elif ty == "float" or ty == "double":
-                tc.assertEqual(round(row1[idx], 5), round(float(row2[idx]), 5))
-            elif ty == "boolean":
-                tc.assertEqual(row1[idx], strtobool(row2[idx]))
-            elif ty == "date":
-                tc.assertEqual(row1[idx], strtodate(row2[idx]))
-            elif ty == "timestamp":
-                tc.assertEqual(
-                    row1[idx].astimezone(timezone.utc).replace(tzinfo=None),
-                    strtots(row2[idx]),
-                )
-            elif ty == "timestamp_ntz":
-                tc.assertEqual(row1[idx], datetime.fromisoformat(row2[idx]))
-            elif ty == "string":
-                tc.assertEqual(row1[idx], row2[idx])
-            elif ty == "decimal":
-                if row2[idx] == "none":
-                    tc.assertTrue(row1[idx] is None)
-                else:
-                    tc.assertEqual(row1[idx], decimal.Decimal(row2[idx]))
-            else:
-                tc.assertEqual(str(row1[idx]), str(row2[idx]))
-
-
-def compare_sql(args, cmp_sqls):
-    assert len(cmp_sqls) == 2
-    spark = get_spark(args)
-    df1 = spark.sql(cmp_sqls[0])
-    df2 = spark.sql(cmp_sqls[1])
-
-    tc = unittest.TestCase()
-    diff_df = df1.exceptAll(df2).collect()
-    log(f"diff {diff_df}", level=LogLevel.INFO)
-    tc.assertEqual(len(diff_df), 0)
-    diff_df = df2.exceptAll(df1).collect()
-    log(f"diff {diff_df}", level=LogLevel.INFO)
-    tc.assertEqual(len(diff_df), 0)
-
-
-def drop_table(args, drop_sqls):
-    spark = get_spark(args)
-    try:
-        for sql in drop_sqls:
-            log(f"Executing sql: {sql}", level=LogLevel.INFO)
-            spark.sql(sql)
-    except Exception as e:
-        log(f"drop table failed: {e}", level=LogLevel.ERROR)
-        for db in spark.catalog.listDatabases():
-            tables = spark.catalog.listTables(db.name)
-            log(f"existing tables: {tables}", level=LogLevel.INFO)
-        raise e
-
-
-def parse_row(row):
-    result = []
-    current = ""
-    parenthesis_count = {"{": 0, "[": 0, "(": 0}
-    for char in row:
-        if char in parenthesis_count:
-            parenthesis_count[char] += 1
-        elif char == "}":
-            parenthesis_count["{"] -= 1
-        elif char == "]":
-            parenthesis_count["["] -= 1
-        elif char == ")":
-            parenthesis_count["("] -= 1
-
-        if char == "," and all(value == 0 for value in parenthesis_count.values()):
-            result.append(current.strip())
-            current = ""
-        else:
-            current += char
-
-    if current:
-        result.append(current.strip())
-
-    return result
-
-
-def discover_test_cases() -> List[str]:
-    """Discover all test case files in the test_case directory."""
-    test_case_dir = os.path.join(os.path.dirname(__file__), "test_case")
-    # Find all .toml files in test_case directory, excluding benches directory
-    test_files = glob.glob(os.path.join(test_case_dir, "*.toml"))
-    test_files.extend(glob.glob(os.path.join(test_case_dir, "*/*.toml")))
-    # Filter out bench files
-    # Hack: filter out CDC tests, as it uses `--host=mysql` (for CI).
-    # TODO: use `risedev slt` to replace `sqllogictest` so that the test can be run locally.
-    test_files = [f for f in test_files if "benches" not in f and "cdc" not in f]
-    return test_files
-
-
-class TestResult(NamedTuple):
-    name: str
-    success: bool
-    duration: timedelta
-    error: Optional[Exception]
-
-
-def run_test_case(test_file: str, args: Dict[str, Any]) -> TestResult:
+def run_test_case(test_file: str, args: dict) -> TestResult:
     """Run a single test case."""
     log(f"\nRunning test case: {test_file}", level=LogLevel.HEADER)
     start_time = datetime.now()
@@ -281,6 +57,13 @@ def run_test_case(test_file: str, args: Dict[str, Any]) -> TestResult:
                 execute_slt(args, verify_slt)
             if drop_sqls:
                 drop_table(args, drop_sqls)
+    except KeyboardInterrupt:
+        log(f"test case {test_file} interrupted by user", level=LogLevel.ERROR)
+        error = "Interrupted by user"
+        success = False
+        # Don't raise the exception, just return the failed result
+        duration = datetime.now() - start_time
+        return TestResult(test_file, success, duration, error)
     except Exception as e:
         log(f"test case {test_file} failed: {e}", level=LogLevel.ERROR)
         error = e
@@ -291,49 +74,6 @@ def run_test_case(test_file: str, args: Dict[str, Any]) -> TestResult:
         return TestResult(test_file, success, duration, error)
 
 
-def print_test_summary(results: List[TestResult]):
-    """Print a summary of all test results."""
-    log("\nTest Summary", level=LogLevel.HEADER)
-    log("", level=LogLevel.SEPARATOR)
-
-    total_duration = timedelta()
-    passed = 0
-    failed = 0
-
-    # Print individual test results
-    for result in results:
-        status = (
-            Color.GREEN + "PASSED" + Color.ENDC
-            if result.success
-            else Color.RED + "FAILED" + Color.ENDC
-        )
-        duration_str = f"{result.duration.total_seconds():.2f}s"
-        log(
-            f"{result.name}: {status} ({duration_str})",
-            level=LogLevel.INFO if result.success else LogLevel.ERROR,
-        )
-        if not result.success and result.error:
-            log(f"  Error: {str(result.error)}", level=LogLevel.ERROR, indent=2)
-
-        total_duration += result.duration
-        if result.success:
-            passed += 1
-        else:
-            failed += 1
-
-    # Print summary statistics
-    log("", level=LogLevel.SEPARATOR)
-    total_tests = len(results)
-    pass_rate = (passed / total_tests * 100) if total_tests > 0 else 0
-
-    log(f"Total Tests: {total_tests}", level=LogLevel.RESULT)
-    log(f"Passed: {passed}", level=LogLevel.RESULT)
-    log(f"Failed: {failed}", level=LogLevel.RESULT)
-    log(f"Pass Rate: {pass_rate:.1f}%", level=LogLevel.RESULT)
-    log(f"Total Duration: {total_duration.total_seconds():.2f}s", level=LogLevel.RESULT)
-    log("", level=LogLevel.SEPARATOR)
-
-
 def get_parallel_job_info() -> Optional[tuple[int, int]]:
     """Get parallel job information from environment variables."""
     job = os.environ.get("BUILDKITE_PARALLEL_JOB")
@@ -342,30 +82,6 @@ def get_parallel_job_info() -> Optional[tuple[int, int]]:
     job = int(job)
     total = int(os.environ.get("BUILDKITE_PARALLEL_JOB_COUNT"))
     return job, total
-
-
-def prepare_test_env():
-    log("prepare test env", level=LogLevel.HEADER)
-    log("create minio bucket", level=LogLevel.INFO)
-    subprocess.run(
-        [
-            "risedev",
-            "mc",
-            "mb",
-            "-p",
-            "hummock-minio/icebergdata",
-        ],
-        check=True,
-    )
-    log("start spark connect server", level=LogLevel.INFO)
-    subprocess.run(
-        [
-            os.path.join(os.path.dirname(__file__), "start_spark_connect_server.sh"),
-        ],
-        check=True,
-    )
-    time.sleep(3)
-    log("prepare test env done", level=LogLevel.HEADER)
 
 
 if __name__ == "__main__":
@@ -410,12 +126,18 @@ if __name__ == "__main__":
 
             # Run all test cases assigned to this job
             for test_file in test_files:
-                try:
-                    result = run_test_case(test_file, config_dict)
-                    test_results.append(result)
-                except Exception:
-                    # Continue running other tests even if one fails
-                    continue
+                result = run_test_case(test_file, config_dict)
+                test_results.append(result)
+                if (
+                    not result.success
+                    and isinstance(result.error, str)
+                    and "Interrupted" in result.error
+                ):
+                    # If the test was interrupted, stop running more tests
+                    break
     finally:
-        # Always print summary, even if some tests failed
+        # Always print summary, even if some tests failed or were interrupted
         print_test_summary(test_results)
+        # Exit with error code if any test failed
+        if any(not result.success for result in test_results):
+            sys.exit(1)
