+++ e2e_test/iceberg/test_utils/__init__.py
@@ -1,0 +1,11 @@
+from .log import LogLevel, log
+from .test_result import TestResult, print_test_summary
+from .test_utils import (
+    init_iceberg_table,
+    compare_sql,
+    drop_table,
+    discover_test_cases,
+    execute_slt,
+    prepare_test_env,
+    verify_result,
+)
